# swagger.yaml
openapi: 3.0.0
info:
  title: Pacific Telecom
  description: Apis para Pacific Telecom
  version: 0.1
servers:
  # - url: http://localhost:3000
  - url: http://138.197.105.129:3000
    description: Pacific Telecom Resource Server

# 1) Define the security scheme type (HTTP bearer)
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
# 2) Apply the security globally to all operations
security:
  - bearerAuth: []         # use the same name as above
paths:
...
  /failure_reports/{id}:
    get:
      summary: Obtener un reporte de falla por id
      tags:
        - Reportes de fallas
      parameters:
        - name: id
          in: path
          required: true
          description: id del reporte de falla
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

    /failure_reports/{id}:
    delete:
      summary: Borra un reporte de falla por id
      tags:
        - Reportes de fallas
      parameters:
        - name: id
          in: path
          required: true
          description: id del reporte de falla
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

      /failure_reports/{id}:
    patch: 
      summary: Actualizar un reporte de falla por id
      tags:
        - Reportes de fallas
      parameters:
        - name: id
          in: path
          required: true
          description: id del reporte de falla
          schema:
            type : integer
            format: int64
            minimum: 1
      requestBody:   
        content:
          application/json:
            schema:
              type: object
            example: '{"subject": "test","description": "test","type":1,"service_id": 1,"client_id":2}'
      responses: 
        '200':
          description: OK
  /failure_reports/register:
    post: 
      summary: Registrar un Reportes de falla
      tags:
        - Reportes de fallas
      requestBody:   
        content:
          application/json:
            schema:
              type: object
            example: '{"subject": "test","description": "test","type":1,"service_id": 1,"client_id":2}'
      responses: 
        '201':
          description: OK

 ############################################### [ ASOCIAR PLANES A USUARIOS ] ############################################
...
  /email_template/{id}:
    get:
      summary: Obtener un template
      tags:
         - Email Templates
      parameters:
        - name: id
          in: path
          required: true
          description: id del template
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK

  
    /email_template/{id}:
    patch: 
      summary: Actualizar template
      tags:
        - Email Templates
      parameters:
        - name: id
          in: path
          required: true
          description:  id del template
          schema:
            type : integer
            format: int64
            minimum: 1
      requestBody:   
        content:
          application/json:
            schema:
              type: object
              properties:
                 name:
                  required: true
                  type: string
                 text:
                  required: true
                  type: text
                 password:
                  required: true
                  type: boolean
            example: '{"name": "test","text":"test" ,"status": true}'
      responses: 
        '200':
          description: OK

      /email_template/{id}:
    delete:
      summary: Borra un template por id
      tags:
        - Email Templates
      parameters:
        - name: id
          in: path
          required: true
          description:  id de la configuracion del servidor
          schema:
            type : integer
            format: int64
            minimum: 1
      responses: 
        '200':
          description: OK
  /email_template/register:
    post: 
      summary: Registrar configuracion del servidor
      tags:
        - Email Templates
      requestBody:   
        content:
          application/json:
            schema:
              type: object
              properties:
                 name:
                  required: true
                  type: string
                 text:
                  required: true
                  type: text
                 password:
                  required: true
                  type: boolean
            example: '{"name": "test","text":"test" ,"status": true}'
      responses: 
        '201':
          description: OK


           ############################################### [CONFIGURACION DE INTERFAZ DE RED] ############################################
...